Move AccountResult to its own file
